refactor template and runner
