+++ src/matcha_ml/templates/__init__.py
@@ -1,1 +1,5 @@
-"""Init for the templates package."""
+"""Matcha templates sub-module."""
+from .azure_template import AzureTemplate
+from .remote_state_template import RemoteStateTemplate
+
+__all__ = ["AzureTemplate", "RemoteStateTemplate"]
