+++ tests/test_cli/test_destroy.py
@@ -51,7 +51,7 @@ def test_cli_destroy_command_with_no_provisioned_resources(
 
     # Invoke destroy command
     with patch(
-        "matcha_ml.templates.azure_template.azure_template.check_current_deployment_exists"
+        "matcha_ml.templates.azure_template.check_current_deployment_exists"
     ) as check_deployment_exists:
         check_deployment_exists.return_value = False
         result = runner.invoke(app, ["destroy"])
