+++ tests/test_runners/test_base_runner.py
@@ -1,0 +1,151 @@
+"""Test for testing BaseRuner class."""
+import os
+from unittest import mock
+from unittest.mock import MagicMock
+
+import pytest
+import typer
+from _pytest.capture import SysCapture
+
+from matcha_ml.errors import MatchaTerraformError
+from matcha_ml.runners.base_runner import BaseRunner
+
+
+CLASS_STUB = "matcha_ml.runners.base_runner.BaseRunner"
+
+
+def test_check_terraform_installation(capsys: SysCapture):
+    """Test app exits when terraform is not installed.
+
+    Args:
+        capsys (SysCapture): fixture to capture stdout and stderr
+    """
+    expected = "Terraform is not installed"
+    runner = BaseRunner()
+    runner.tfs = MagicMock()
+    runner.tfs.check_installation.return_value = False
+    with pytest.raises(typer.Exit):
+        runner._check_terraform_installation()
+    captured = capsys.readouterr()
+
+    assert expected in captured.err
+
+
+def test_validate_terraform_config(capsys: SysCapture):
+    """Test application exits if there is no config.
+
+    Args:
+        capsys (SysCapture): fixture to capture stdout and stderr
+    """
+    expected = "The file terraform.tfvars.json was not found"
+
+    runner = BaseRunner()
+    runner.tfs = MagicMock()
+    runner.tfs.validate_config.return_value = False
+
+    with pytest.raises(typer.Exit):
+        runner._validate_terraform_config()
+    captured = capsys.readouterr()
+
+    assert expected in captured.err
+
+
+# def test_is_approved(template_runner: BaseRunner):
+#     """Test if is_approved behaves as expected based on user's input.
+
+#     Args:
+#         template_runner (BaseRunner): a BaseRunner object instance
+#     """
+#     with mock.patch("typer.confirm") as mock_confirm:
+#         mock_confirm.return_value = True
+#         assert template_runner.is_approved("provision")
+
+#         mock_confirm.return_value = False
+#         assert not template_runner.is_approved("provision")
+
+
+# def test_initialize_terraform(capsys: SysCapture, template_runner: BaseRunner):
+#     """Test if service behaves as expected when initializing Terraform.
+
+#     Args:
+#         capsys (SysCapture): fixture to capture stdout and stderr
+#         template_runner (BaseRunner): a BaseRunner object instance
+#     """
+#     template_runner.tf_state_dir = MagicMock()
+
+#     with mock.patch.object(template_runner.tf_state_dir, "exists", return_value=True):
+#         expected = "has already been initialized"
+
+#         template_runner._initialize_terraform()
+
+#         captured = capsys.readouterr()
+
+#         assert expected in captured.out
+
+#     with mock.patch.object(template_runner.tf_state_dir, "exists", return_value=False):
+#         template_runner.tfs.init = MagicMock(return_value=(0, "", ""))
+#         expected = " initialized!"
+#         template_runner._initialize_terraform()
+
+#         captured = capsys.readouterr()
+
+#         assert expected in captured.out
+
+
+# def test_check_matcha_directory_exists(
+#     capsys: SysCapture,
+#     template_runner: BaseRunner,
+#     matcha_testing_directory: str,
+# ):
+#     """Test if service exit as expected and print out the expected error message when required files does not exists.
+
+#     Args:
+#         capsys (SysCapture): fixture to capture stdout and stderr
+#         template_runner (BaseRunner): a BaseRunner object instance
+#         matcha_testing_directory (str): the test directory
+#     """
+#     os.chdir(matcha_testing_directory)
+
+#     template_runner.tfs.check_matcha_directory_exists = MagicMock(return_value=False)
+#     template_runner.tfs.check_matcha_directory_integrity = MagicMock(return_value=False)
+
+#     with pytest.raises(typer.Exit):
+#         expected = "Error, the .matcha directory does not exist"
+#         template_runner._check_matcha_directory_exists()
+
+#         captured = capsys.readouterr()
+
+#         assert expected in captured.err
+
+#     with pytest.raises(typer.Exit):
+#         expected = "Error, the .matcha directory does not contain files relating to deployed resources. Please ensure you are trying to destroy resources that you have provisioned in the current working directory."
+#         template_runner._check_matcha_directory_exists()
+
+#         captured = capsys.readouterr()
+
+#         assert expected in captured.err
+
+
+# def test_apply_terraform(capsys: SysCapture, template_runner: BaseRunner):
+#     """Test if terraform applied is handled correctly during apply when provisioning resources.
+
+#     Args:
+#         capsys (SysCapture): fixture to capture stdout and stderr
+#         template_runner (BaseRunner): a BaseRunner object instance
+#     """
+#     template_runner.tfs.apply = MagicMock(return_value=(0, "", ""))
+#     expected = "Your environment has been provisioned!"
+
+#     template_runner._apply_terraform()
+#     captured = capsys.readouterr()
+
+#     assert expected in captured.out
+
+#     template_runner.tfs.apply = MagicMock(return_value=(1, "", "Apply failed"))
+
+#     with pytest.raises(MatchaTerraformError) as exc_info:
+#         template_runner._apply_terraform()
+#         assert (
+#             str(exc_info.value)
+#             == "Terraform failed because of the following error: 'Apply failed'."
+#         )
