+++ tests/test_state/test_remote_state_manager.py
@@ -17,12 +17,8 @@ from matcha_ml.state.remote_state_manager import (
     RemoteStateConfig,
     RemoteStateManager,
 )
-from matcha_ml.templates.state_storage_template.run_state_storage_template import (
-    StateStorageTemplateRunner,
-)
-from matcha_ml.templates.state_storage_template.state_storage_template import (
-    SUBMODULE_NAMES,
-)
+from matcha_ml.runners import RemoteStateRunner
+from matcha_ml.templates.remote_state_template import SUBMODULE_NAMES
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 TEMPLATE_DIR = os.path.join(
@@ -214,7 +210,7 @@ def test_deprovision_state_storage(capsys: SysCapture) -> None:
         capsys (SysCapture): fixture to capture stdout and stderr
     """
     with patch(
-        "matcha_ml.templates.state_storage_template.run_state_storage_template.StateStorageTemplateRunner.deprovision"
+        "matcha_ml.runners.remote_state_runner.RemoteStateRunner.deprovision"
     ) as destroy:
         destroy.return_value = None
         remote_state_manager = RemoteStateManager()
@@ -225,7 +221,7 @@ def test_deprovision_state_storage(capsys: SysCapture) -> None:
 
         expected_output = "Destroying remote state management is complete!"
 
-        template_runner = StateStorageTemplateRunner()
+        template_runner = RemoteStateRunner()
         template_runner.deprovision.assert_called()
 
         assert expected_output in captured.out
